Add AppStream metadata
This makes the tool show up in GNOME Software and, more importantly, be
offered when the user plugs in the hardware (with the modalias match).

This makes it possible for a Linux user to trivially discover the right
tool when they plug in the hardware.

Again, not really integrated with the project file, so the packager is
responsible for copying this to the right place.